+++ static/plugins/karel/runner/karel.worker.js
@@ -169,6 +169,9 @@ class KarelAPI extends BaseAPI {
 let api;
 
 const onAnyMessage = async (event) => {
+  // Whether or not a fake prompt + command is echoed on the terminal
+  if (api) api.echoCmd = event.data.data?.echoCmd !== false;
+
   switch (event.data.id) {
     case 'constructor': {
       const { port } = event.data.data;